linux-adk renamed to adk
