+++ simple-rt-cli/include/adk.h
